refactor(DataFlag): re-simplify by removing baseclass
