+++ chimedb/dataflag/orm.py
@@ -8,7 +8,7 @@ from future.builtins.disabled import *  # noqa  pylint: disable=W0401, W0614
 
 # === End Python 2/3 compatibility
 
-from .base_models import DataSubset, DataSubsetType
+from .base_models import DataSubsetType
 
 from chimedb.core.orm import base_model, JSONDictField, EnumField
 from chimedb.core.exceptions import ValidationError
@@ -69,7 +69,7 @@ class DataFlagType(DataSubsetType):
     pass
 
 
-class DataFlag(DataSubset):
+class DataFlag(base_model):
     """A flagged range of data.
 
     Attributes
@@ -102,6 +102,11 @@ class DataFlag(DataSubset):
     """
 
     type = pw.ForeignKeyField(DataFlagType, backref="flags")
+
+    start_time = pw.DoubleField()
+    finish_time = pw.DoubleField(null=True)
+
+    metadata = JSONDictField(null=True)
 
     @classmethod
     def create_flag(
@@ -166,6 +171,58 @@ class DataFlag(DataSubset):
             metadata=table_metadata,
         )
         return flag
+
+    @property
+    def instrument(self):
+        """The instrument the flag applies to."""
+        if self.metadata is not None:
+            return self.metadata.get("instrument", None)
+        else:
+            return None
+
+    @property
+    def freq(self):
+        """The list of inputs the flag applies to. `None` if not set."""
+        if self.metadata is not None:
+            return self.metadata.get("freq", None)
+        else:
+            return None
+
+    @property
+    def freq_mask(self):
+        """An array for the frequencies flagged (`True` if the flag applies)."""
+
+        # TODO: hard coded for CHIME
+        mask = np.ones(1024, dtype=np.bool)
+
+        if self.freq is not None:
+            mask[self.freq] = False
+            mask = ~mask
+
+        return mask
+
+    @property
+    def inputs(self):
+        """The list of inputs the flag applies to. `None` if not set."""
+        if self.metadata is not None:
+            return self.metadata.get("inputs", None)
+        else:
+            return None
+
+    @property
+    def input_mask(self):
+        """An array for the inputs flagged (`True` if the flag applies)."""
+        if self.instrument is None:
+            return None
+
+        inp_dict = {"chime": 2048, "pathfinder": 256}
+        mask = np.ones(inp_dict[self.instrument], dtype=np.bool)
+
+        if self.inputs is not None:
+            mask[self.inputs] = False
+            mask = ~mask
+
+        return mask
 
 
 # Tables pertaining to the data flag opinions.
